refactor datagrid cell mutation observer to be much more specific at what it is watching
